Implement CRUD operations for flashcards, inventory, items, pets, and users in the pet_service API
- Added endpoints for inserting and retrieving flashcards.
- Implemented inventory purchase functionality with user coin validation.
- Created item insertion and retrieval endpoints, categorized by type.
- Developed pet retrieval and creation endpoints.
- Established user insertion and answer handling for flashcards, including streak and coin management.
- Enhanced error handling and validation using Pydantic models.
