+++ pet_service/routers/__init__.py
@@ -1,0 +1,3 @@
+from . import flashcards, items, users, inventory, pets
+
+__all__ = ["flashcards", "items", "users", "inventory", "pets"]
