+++ pet_service/test/insert_items.py
@@ -3,9 +3,9 @@ import requests
 url = "http://127.0.0.1:8000/items/insert"
 
 items = [
-        {"itemId": 1, "name": "Smores", "itemType": "food", "price": 20},
-        {"itemId": 2, "name": "Football", "itemType": "toy", "price": 50},
-        {"itemId": 3, "name": "Blue Hoodie", "itemType": "clothes", "price": 100},
+        {"name": "Pizza", "itemType": "FOOD", "price": 10},
+        {"name": "Octopus Plush", "itemType": "TOYS", "price": 70},
+        {"name": "Sunglasses", "itemType": "CLOTHING", "price": 30},
     ]
 
 response = requests.post(url, json=items)
