+++ pet_service/test/test_get_inventory.py
@@ -3,7 +3,7 @@ import requests
 BASE = "http://127.0.0.1:8000"
 
 # Only test the GET /inventory endpoint. Set userId to an existing user in your DB.
-userId = "Chetan"
+userId = "68e1f779dc6d82796f3b3f45"
 
 url = f"{BASE}/inventory?userId={userId}"
 resp = requests.get(url)
