+++ pet_service/test/test_insert_flashcards.py
@@ -3,25 +3,11 @@ import requests
 url = "http://127.0.0.1:8000/flashcards/insert"
 data = [
     {
-        "question": "Who is the coolest guy?",
-        "options": ["Priyal", "Chetan", "Christo"],
+        "question": "What's the capital of France?",
+        "options": ["Berlin", "Paris", "London"],
         "topicId": "12345",
         "spaceId": "67890",
         "correctOption": 0
-    },
-    {
-        "question": "What table has the coolest team at HackRU?",
-        "options": ["Table 99", "Table 67", "Table 24"],
-        "topicId": "54321",
-        "spaceId": "09876",
-        "correctOption": 2
-    },
-    {
-        "question": "What is 2 + 2?",
-        "options": ["3", "4", "5"],
-        "topicId": "11111",
-        "spaceId": "22222",
-        "correctOption": 1
     }
 ]
