+++ pet_service/test/test_insert_user.py
@@ -1,7 +1,7 @@
 import requests
 
 url = "http://127.0.0.1:8000/users/insert"
-data = { "userId": "Chetan" }
+data = { "userId": "Vishal" }
 
 
 response = requests.post(url, json=data)
