+++ pet_service/test/test_users_answer.py
@@ -6,7 +6,7 @@ BASE = "http://127.0.0.1:8000"
 # already exist in your database before running the test.
 answer_url = f"{BASE}/users/answer"
 # adjust these values to match existing records in your DB
-payload = {"userId": "68e1f774023ee7972dddcad1", "flashcardId": "68e1f70845970cfe0cc04fce", "optionSelected": 2}
+payload = {"userId": "68e218483f112a5967f2bbe9", "flashcardId": "68e20b148b78b99940b3892c", "optionSelected": 0}
 
 resp = requests.post(answer_url, json=payload)
 print(resp.status_code)
